requirements.txt
streamlit>=1.45.0
pandas==2.1.4
numpy==1.26.3
matplotlib==3.8.2
plotly==5.19.0
seaborn==0.13.1 
pyyaml